Go refactor
