+++ 2023/day_01/task_1/main.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package main
 
 import (
