+++ 2023/day_01/task_1/wasm.go
@@ -1,22 +1,16 @@
+//go:build js
+
 package main
 
 import (
-	"unsafe"
+	"syscall/js"
 )
 
-// We store the result as a global variable to prevent it being
-// lost to GC before the caller can read it
-var result []byte
+func main() {
+	js.Global().Set("solve", js.FuncOf(func(this js.Value, args []js.Value) interface{} {
+		input := args[0].String()
+		return Solve(input)
+	}))
 
-//export solve
-func solve(ptr uint32, size uint32) uint64 {
-	inputBytes := unsafe.Slice((*byte)(unsafe.Pointer(uintptr(ptr))), size)
-	input := string(inputBytes)
-
-	result = []byte(Solve(input))
-
-	resultPtr := uintptr(unsafe.Pointer(&result[0]))
-	resultLen := uint64(len(result))
-
-	return (uint64(resultPtr) << 32) | resultLen
+	select {}
 }
